docs(auto-focus): move auto-focus stories to chromatic
