+++ src/components/date/date.mdx
@@ -53,10 +53,6 @@ be used to set the input value like in the example below, although the component
 
 <Canvas of={DateStories.Sizes} />
 
-### AutoFocus
-
-<Canvas of={DateStories.AutoFocus} />
-
 ### Disabled
 
 <Canvas of={DateStories.Disabled} />
@@ -117,7 +113,7 @@ but the date format has been overridden to `yyyy-M-d` as opposed to the default 
 
 ### With labelInline (legacy)
 
-Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and 
+Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and
 `labelAlign` props to control the width and text alignment of the label and the `inputWidth` prop to control the width of the input.
 
 **Note:** This is a legacy feature and will only render if the `validationRedesignOptIn` feature flag on an ancestor [CarbonProvider](../?path=/docs/carbon-provider--docs) is *false*.
