+++ src/components/number/number.mdx
@@ -46,17 +46,13 @@ before any other announcements, this well help screen reader users understand th
 
 <Canvas of={NumberStories.ReadOnly} />
 
-### AutoFocus
-
-<Canvas of={NumberStories.AutoFocus} />
-
 ### With custom maxWidth
 
 <Canvas of={NumberStories.WithCustomMaxWidth} />
 
 ### With labelInline (legacy)
 
-Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and 
+Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and
 `labelAlign` props to control the width and text alignment of the label and the `inputWidth` prop to control the width of the input.
 
 **Note:** This is a legacy feature and will only render if the `validationRedesignOptIn` feature flag on an ancestor [CarbonProvider](../?path=/docs/carbon-provider--docs) is *false*.
