+++ src/components/password/password.mdx
@@ -82,10 +82,6 @@ before any other announcements, this well help screen reader users understand th
 
 <Canvas of={PasswordStories.ReadOnly} />
 
-### AutoFocus
-
-<Canvas of={PasswordStories.AutoFocus} />
-
 ### With custom maxWidth
 
 <Canvas of={PasswordStories.WithCustomMaxWidth} />
@@ -96,7 +92,7 @@ before any other announcements, this well help screen reader users understand th
 
 ### With labelInline (legacy)
 
-Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and 
+Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and
 `labelAlign` props to control the width and text alignment of the label and the `inputWidth` prop to control the width of the input.
 
 **Note:** This is a legacy feature and will only render if the `validationRedesignOptIn` feature flag on an ancestor [CarbonProvider](../?path=/docs/carbon-provider--docs) is *false*.
