+++ src/components/textarea/textarea.mdx
@@ -48,10 +48,6 @@ import Textarea from "carbon-react/lib/components/textarea";
 
 <Canvas of={TextareaStories.ReadOnlyStory} />
 
-### AutoFocus
-
-<Canvas of={TextareaStories.AutoFocusStory} />
-
 ### Expandable
 
 <Canvas of={TextareaStories.ExpandableStory} />
@@ -112,7 +108,7 @@ If you want to remove the borders from the component, you can use the `hideBorde
 
 ### With labelInline (legacy)
 
-Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and 
+Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and
 `labelAlign` props to control the width and text alignment of the label and the `inputWidth` prop to control the width of the input.
 
 **Note:** This is a legacy feature and will only render if the `validationRedesignOptIn` feature flag on an ancestor [CarbonProvider](../?path=/docs/carbon-provider--docs) is *false*.
