+++ src/components/textbox/textbox.mdx
@@ -83,10 +83,6 @@ Include the formatted number count wherever makes sense for the language you're 
 
 <Canvas of={TextboxStories.Margins} />
 
-### AutoFocus
-
-<Canvas of={TextboxStories.AutoFocus} />
-
 ### Disabled
 
 <Canvas of={TextboxStories.Disabled} />
@@ -113,7 +109,7 @@ You can use the `isOptional` prop to indicate if the field is mandatory.
 
 ### With labelInline (legacy)
 
-Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and 
+Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and
 `labelAlign` props to control the width and text alignment of the label and the `inputWidth` prop to control the width of the input.
 
 **Note:** This is a legacy feature and will only render if the `validationRedesignOptIn` feature flag on an ancestor [CarbonProvider](../?path=/docs/carbon-provider--docs) is *false*.
